Currently when we need to check if a repo is current, we shell out twice using the `sh` module, once to check the current ref, once to do a dry-run of `git fetch`. Both of these are slow and mean it takes a long time to see when the repo is out of date, once it's built.

The net effect of this is that, once you've built everything, your next `.venv/bin/etl --dry-run` spends several minutes just trying to work out what's out of date.

This change switches to getting the local sha by just reading from disk, and doing a single HTTPS request using Git's smart HTTPS protocol to Github to get the latest remote sha.

## `repo.is_up_to_date()` / 1.9x speedup

Before:

```
In [3]: %timeit git.GithubRepo('open-numbers', 'ddf--ihme--global_burden_disease_etiology').is_up_to_date()
428 ms ± 8.22 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

After:

```
In [6]: %timeit git.GithubRepo('open-numbers', 'ddf--ihme--global_burden_disease_etiology').is_up_to_date()
223 ms ± 15.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

## `.venv/bin/etl --dry-run` / 3.5x speedup

On a fully-built ETL, meaning we have c. 100 repositories to check for updates.

```
➜  etl git:(master) time .venv/bin/etl --dry-run
Detecting which steps need rebuilding...
Running 2 steps:
1. github://open-numbers/ddf--undp--hdi...
2. data://meadow/open_numbers/latest/undp__hdi...
.venv/bin/etl --dry-run  7.97s user 7.24s system 9% cpu 2:39.30 total
```

After:

```
➜  etl git:(master) time .venv/bin/etl --dry-run
Detecting which steps need rebuilding...
Running 2 steps:
1. github://open-numbers/ddf--undp--hdi...
2. data://meadow/open_numbers/latest/undp__hdi...
.venv/bin/etl --dry-run  4.53s user 2.35s system 15% cpu 45.378 total
```