@larsyencken 2021-11-11
Thank you both!

@danyx23 So far I'm trying to avoid the number of credentials you need to work with it, which is why I've avoided using the API directly (that was an earlier implementation). We could totally go back there if it gets problematic. Likewise we could find a way to do these checks async instead of sequentially. But... that's another kettle of fish 😅